widget form refactoring name and subdir
